Use v7 insert on whmcs 8
We should use the same insert to the database for whmcs version 8 that the one we use on whmcs > 7.6.
Previous if would make use the older method, as for now dd_get_whmcs_sub_version() is null. 
Can also be aproached by making the check for it bigger than 7, so that it can be available for newer versions, but this can also break those newer versions that haven't been compatibility checked.